requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.36
pydantic==2.9.0
python-multipart==0.0.12
python-dotenv==1.0.1
alembic==1.14.0
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
bcrypt==4.2.0